+++ API/DTOs/OrderItemDto.cs
@@ -11,7 +11,7 @@ namespace API.DTOs
         public string Name { get; set; }
 
         public string PictureUrl { get; set; }
-        public decimal Price  { get; set; }
+        public long Price  { get; set; }
 
         public int Quantity { get; set; }
